pyproject.toml
[tool.poetry]
name = "runflow-docs"
version = "0.1.0"
description = ""
authors = ["Seungmin <loglog354@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.54.1"
pathlib = "^1.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
